+++ docs/md/releasenotes.md
@@ -5,6 +5,8 @@
 - Enable fix previously behind system property: fix `TypedReferenceValue.generalize()` not setting `mayBeExtension` to true when generalizing to common parent type.
 - Avoid printing `PartialEvaluator` messages when an `ExcessiveComplexityException` occurs.
 - Fix incorrect writing of flags for type parameters with name annotations.
+- Fix incorrect writing of flags for reified type parameters.
+- Fix model for types and type parameters, removing the incorrect `HAS_ANNOTATION` common flag.
 
 ### Improved